Crate poly_iop

Macros

to_bytes

Takes as input a struct, and converts them to a series of bytes. All traits that implement CanonicalSerialize can be automatically converted to bytes in this manner.

Structs

IOPTranscript

An IOP transcript consists of a Merlin transcript and a flag is_empty to indicate that if the transcript is empty.

PolyIOP

Struct for PolyIOP protocol. It has an associated type F that defines the prime field the multi-variate polynomial operates on.

VPAuxInfo

Auxiliary information about the multilinear polynomial

VirtualPolynomial

A virtual polynomial is a sum of products of multilinear polynomials; where the multilinear polynomials are stored via their multilinear extensions: (coefficient, DenseMultilinearExtension)

Enums

PolyIOPErrors

A enum specifying the possible failure modes of the PolyIOP.

Traits

HyperPlonkPIOP

A trait for HyperPlonk Poly-IOPs

PermutationCheck

A PermutationCheck is derived from ZeroCheck.

ProductCheck

A ProductCheck is derived from ZeroCheck.

SumCheck

Trait for doing sum check protocols.

ZeroCheck

A ZeroCheck is derived from SumCheck.

Functions

bit_decompose

Decompose an integer into a binary vector in little endian.

identity_permutation_mle

An MLE that represent an identity permutation: f(index) \mapto index

random_permutation_mle

An MLE that represent a random permutation