Module poly_iop::prelude

Re-exports

pub use crate::PolyIOP;

Macros

to_bytes

Takes as input a struct, and converts them to a series of bytes. All traits that implement CanonicalSerialize can be automatically converted to bytes in this manner.

Enums

PolyIOPErrors

A enum specifying the possible failure modes of the PolyIOP.

Traits

PermutationCheck

A PermutationCheck is derived from ZeroCheck.

SumCheck

Trait for doing sum check protocols.

ZeroCheck

A ZeroCheck is derived from SumCheck.

Functions

bit_decompose

Decompose an integer into a binary vector in little endian.

identity_permutation_mle

An MLE that represent an identity permutation: f(index) \mapto index

random_permutation_mle

An MLE that represent a random permutation