Trait poly_iop::prelude::PermutationCheck

pub trait PermutationCheck<F: PrimeField>: ZeroCheck<F> {
    type PermutationCheckSubClaim;
    type PermutationChallenge;
    type PermutationProof;

    fn preprocess(
        permutation: &[F], 
        aux_info: &Self::VPAuxInfo
    ) -> Result<DenseMultilinearExtension<F>, PolyIOPErrors>;
    fn init_transcript() -> Self::Transcript;
    fn generate_challenge(
        transcript: &mut Self::Transcript
    ) -> Result<Self::PermutationChallenge, PolyIOPErrors>;
    fn compute_prod_evals(
        challenge: &Self::PermutationChallenge, 
        fx: &DenseMultilinearExtension<F>, 
        gx: &DenseMultilinearExtension<F>, 
        s_perm: &DenseMultilinearExtension<F>
    ) -> Result<[DenseMultilinearExtension<F>; 3], PolyIOPErrors>;
    fn update_challenge(
        challenge: &mut Self::PermutationChallenge, 
        transcript: &mut Self::Transcript, 
        prod_x_binding: &impl CanonicalSerialize
    ) -> Result<(), PolyIOPErrors>;
    fn prove(
        prod_x_and_aux_info: &[DenseMultilinearExtension<F>; 3], 
        challenge: &Self::PermutationChallenge, 
        transcript: &mut IOPTranscript<F>
    ) -> Result<Self::PermutationProof, PolyIOPErrors>;
    fn verify(
        proof: &Self::PermutationProof, 
        aux_info: &Self::VPAuxInfo, 
        transcript: &mut Self::Transcript
    ) -> Result<Self::PermutationCheckSubClaim, PolyIOPErrors>;

    fn commit_prod_x() { ... }
}

A PermutationCheck is derived from ZeroCheck.

A Permutation Check IOP takes the following steps:

Inputs:

f(x)

g(x)

permutation s_perm(x)

Steps:

generate_challenge from current transcript (generate beta, gamma)

compute_product to build prod(x) etc. from f, g and s_perm

push a commitment of prod(x) to the transcript (done by the snark caller)

update_challenge with the updated transcript (generate alpha)

prove to generate the proof

Required Associated Types

type PermutationCheckSubClaim

type PermutationChallenge

type PermutationProof

Required Methods

fn preprocess(
permutation: &[F],
aux_info: &Self::VPAuxInfo
) -> Result<DenseMultilinearExtension<F>, PolyIOPErrors>

Generate the preprocessed polynomial for the permutation check.

The algorithm takes as input a permutation and outputs a merged multilinear polynomial s(X0, X1, …, Xn) such that

s(0, X1, …, Xn) = s_id(X1, …, Xn) (identity permutation polynomial)

s(1, X1, …, Xn) = s_perm(X1, …, Xn) (permutation polynomial)

fn init_transcript() -> Self::Transcript

Initialize the system with a transcript

This function is optional – in the case where a PermutationCheck is an building block for a more complex protocol, the transcript may be initialized by this complex protocol, and passed to the PermutationCheck prover/verifier.

fn generate_challenge(
transcript: &mut Self::Transcript
) -> Result<Self::PermutationChallenge, PolyIOPErrors>

Step 1 of the IOP. Generate challenge beta and gamma from a transcript.

fn compute_prod_evals(
challenge: &Self::PermutationChallenge,
fx: &DenseMultilinearExtension<F>,
gx: &DenseMultilinearExtension<F>,
s_perm: &DenseMultilinearExtension<F>
) -> Result<[DenseMultilinearExtension<F>; 3], PolyIOPErrors>

Step 2 of the IOP.

Input:

f(x), g(x), s_perm(x) are mle-s

challenges, consists of beta and gamma

Output: the evaluations for the following 3 polynomials

prod(x)

numerator

denominator

where

prod(0,x) := prod(0, x0, x1, …, xn) which is the MLE over the evaluations of the following polynomial on the boolean hypercube {0,1}^n:

(f(x) + \beta s_id(x) + \gamma)/(g(x) + \beta s_perm(x) + \gamma)

where s_id(x) is an identity permutation

numerator is the MLE for f(x) + \beta s_id(x) + \gamma

denominator is the MLE for g(x) + \beta s_perm(x) + \gamma

The caller needs to check num_vars matches in f/g/s_id/s_perm Cost: linear in N.

TODO: replace argument s_perm with the merged polynomial s.

fn update_challenge(
challenge: &mut Self::PermutationChallenge,
transcript: &mut Self::Transcript,
prod_x_binding: &impl CanonicalSerialize
) -> Result<(), PolyIOPErrors>

Step 4 of the IOP. Update the challenge with alpha; returns an error if alpha already exists.

fn prove(
prod_x_and_aux_info: &[DenseMultilinearExtension<F>; 3],
challenge: &Self::PermutationChallenge,
transcript: &mut IOPTranscript<F>
) -> Result<Self::PermutationProof, PolyIOPErrors>

Step 5 of the IOP.

Initialize the prover to argue that an MLE g(x) is a permutation of MLE f(x) over a permutation given by s_perm. Inputs:

3 MLEs from the second step

challenge: Self::Challenge that has been updated

transcript: a transcript that is used to generate the challenges beta and gamma Cost: O(N)

fn verify(
proof: &Self::PermutationProof,
aux_info: &Self::VPAuxInfo,
transcript: &mut Self::Transcript
) -> Result<Self::PermutationCheckSubClaim, PolyIOPErrors>

Verify that an MLE g(x) is a permutation of MLE f(x) over a permutation given by s_perm.

Provided Methods

fn commit_prod_x()

Step 3 of the IOP. push a commitment of prod(x) to the transcript IMPORTANT: this step is done by the snark caller