Trait poly_iop::prelude::ProductCheck

pub trait ProductCheck<E, PCS>: ZeroCheck<E::Fr> where
    E: PairingEngine,
    PCS: PolynomialCommitmentScheme<E>, {
    type ProductCheckSubClaim;
    type ProductCheckProof;

    fn init_transcript() -> Self::Transcript;
    fn prove(
        pcs_param: &PCS::ProverParam, 
        fx: &Self::MultilinearExtension, 
        gx: &Self::MultilinearExtension, 
        transcript: &mut IOPTranscript<E::Fr>
    ) -> Result<(Self::ProductCheckProof, Self::MultilinearExtension), PolyIOPErrors>;
    fn verify(
        proof: &Self::ProductCheckProof, 
        aux_info: &VPAuxInfo<E::Fr>, 
        transcript: &mut Self::Transcript
    ) -> Result<Self::ProductCheckSubClaim, PolyIOPErrors>;
}

A product-check proves that two n-variate multilinear polynomials f(x), g(x) satisfy: \prod_{x \in {0,1}^n} f(x) = \prod_{x \in {0,1}^n} g(x)

A ProductCheck is derived from ZeroCheck.

Prover steps:

build prod(x0, ..., x_n) from f and g, such that prod(0, x1, ..., xn) equals f/g over domain {0,1}^n

push commitments of prod(x) to the transcript, and generate_challenge from current transcript (generate alpha)

generate the zerocheck proof for the virtual polynomial prod(1, x) - prod(x, 0) * prod(x, 1) + alpha * (f(x) - prod(0, x) * g(x))

Verifier steps:

Extract commitments of prod(x) from the proof, push them to the transcript

generate_challenge from current transcript (generate alpha)

verify to verify the zerocheck proof and generate the subclaim for polynomial evaluations

Required Associated Types

type ProductCheckSubClaim

type ProductCheckProof

Required Methods

fn init_transcript() -> Self::Transcript

Initialize the system with a transcript

This function is optional – in the case where a ProductCheck is an building block for a more complex protocol, the transcript may be initialized by this complex protocol, and passed to the ProductCheck prover/verifier.

fn prove(
pcs_param: &PCS::ProverParam,
fx: &Self::MultilinearExtension,
gx: &Self::MultilinearExtension,
transcript: &mut IOPTranscript<E::Fr>
) -> Result<(Self::ProductCheckProof, Self::MultilinearExtension), PolyIOPErrors>

Generate a proof for product check, showing that witness multilinear polynomials f(x), g(x) satisfy \prod_{x \in {0,1}^n} f(x) = \prod_{x \in {0,1}^n} g(x)

Inputs:

fx: the numerator multilinear polynomial

gx: the denominator multilinear polynomial

transcript: the IOP transcript

pk: PCS committing key

Outputs

the product check proof

the product polynomial (used for testing)

Cost: O(N)

fn verify(
proof: &Self::ProductCheckProof,
aux_info: &VPAuxInfo<E::Fr>,
transcript: &mut Self::Transcript
) -> Result<Self::ProductCheckSubClaim, PolyIOPErrors>

Verify that for witness multilinear polynomials f(x), g(x) it holds that \prod_{x \in {0,1}^n} f(x) = \prod_{x \in {0,1}^n} g(x)