Struct poly_iop::PolyIOP

pub struct PolyIOP<F: PrimeField> { /* private fields */ }

Struct for PolyIOP protocol. It has an associated type F that defines the prime field the multi-variate polynomial operates on.

An PolyIOP may be instantiated with one of the following:

SumCheck protocol.

ZeroCheck protocol.

PermutationCheck protocol.

Those individual protocol may have similar or identical APIs. The systematic way to invoke specific protocol is, for example <PolyIOP<F> as SumCheck<F>>::prove()