Trait poly_iop::HyperPlonkPIOP

pub trait HyperPlonkPIOP<F: PrimeField> {
    type Parameters;
    type ProvingKey;
    type Proof;
    type SubClaim;

    fn preprocess(
        params: &Self::Parameters, 
        permutation: &[F], 
        selectors: &[&[F]]
    ) -> Result<Self::ProvingKey, PolyIOPErrors>;
    fn prove(
        pk: &Self::ProvingKey, 
        pub_input: &[F], 
        witness: &[&[F]], 
        transcript: &mut IOPTranscript<F>
    ) -> Result<Self::Proof, PolyIOPErrors>;
    fn verify(
        params: &Self::Parameters, 
        pub_input: &[F], 
        proof: &Self::Proof, 
        transcript: &mut IOPTranscript<F>
    ) -> Result<Self::SubClaim, PolyIOPErrors>;
}

A trait for HyperPlonk Poly-IOPs

Required Associated Types

type Parameters

type ProvingKey

type Proof

type SubClaim

Required Methods

fn preprocess(
params: &Self::Parameters,
permutation: &[F],
selectors: &[&[F]]
) -> Result<Self::ProvingKey, PolyIOPErrors>

Generate the preprocessed polynomials output by the indexer.

Inputs:

params: HyperPlonk instance parameters

permutation: the permutation for the copy constraints

selectors: the list of selector vectors for custom gates Outputs:

The HyperPlonk proving key, which includes the preprocessed polynomials.

fn prove(
pk: &Self::ProvingKey,
pub_input: &[F],
witness: &[&[F]],
transcript: &mut IOPTranscript<F>
) -> Result<Self::Proof, PolyIOPErrors>

Generate HyperPlonk PIOP proof.

Inputs:

pk: circuit proving key

pub_input: online public input

witness: witness assignement

transcript: the transcript used for generating pseudorandom challenges Outputs:

The HyperPlonk PIOP proof.

fn verify(
params: &Self::Parameters,
pub_input: &[F],
proof: &Self::Proof,
transcript: &mut IOPTranscript<F>
) -> Result<Self::SubClaim, PolyIOPErrors>

Verify the HyperPlonk proof and generate the evaluation subclaims to be checked later by the SNARK verifier.

Inputs:

params: instance parameter

pub_input: online public input

proof: HyperPlonk PIOP proof

transcript: the transcript used for generating pseudorandom challenges Outputs:

Return error if the verification fails, otherwise return the evaluation subclaim