Trait poly_iop::PermutationCheck

pub trait PermutationCheck<F: PrimeField>: ZeroCheck<F> {
    type PermutationCheckSubClaim;
    type PermutationChallenge;

    fn init_transcript() -> Self::Transcript;
    fn generate_challenge(
        transcript: &mut Self::Transcript
    ) -> Result<Self::PermutationChallenge, PolyIOPErrors>;
    fn update_challenge(
        challenge: &mut Self::PermutationChallenge, 
        transcript: &mut Self::Transcript, 
        prod_x_binding: &F
    ) -> Result<(), PolyIOPErrors>;
    fn compute_products(
        challenge: &Self::PermutationChallenge, 
        fx: &DenseMultilinearExtension<F>, 
        gx: &DenseMultilinearExtension<F>, 
        s_perm: &DenseMultilinearExtension<F>
    ) -> Result<[DenseMultilinearExtension<F>; 7], PolyIOPErrors>;
    fn prove(
        prod_x_and_aux_info: &[DenseMultilinearExtension<F>; 7], 
        challenge: &Self::PermutationChallenge, 
        transcript: &mut IOPTranscript<F>
    ) -> Result<Self::Proof, PolyIOPErrors>;
    fn verify(
        proof: &Self::Proof, 
        aux_info: &Self::VPAuxInfo, 
        transcript: &mut Self::Transcript
    ) -> Result<Self::PermutationCheckSubClaim, PolyIOPErrors>;
}

A PermutationCheck is derived from ZeroCheck.

A Permutation Check IOP takes the following steps:

Inputs:

f(x)

g(x)

permutation s_perm(x)

Steps:

generate_challenge from current transcript (generate beta, gamma)

compute_product to build prod(x) etc. from f, g and s_perm

push a commitment of prod(x) to the transcript (done by the snark caller)

update_challenge with the updated transcript (generate alpha)

prove to generate the proof

Required Associated Types

type PermutationCheckSubClaim

type PermutationChallenge

Required Methods

fn init_transcript() -> Self::Transcript

Initialize the system with a transcript

This function is optional – in the case where a PermutationCheck is an building block for a more complex protocol, the transcript may be initialized by this complex protocol, and passed to the PermutationCheck prover/verifier.

fn generate_challenge(
transcript: &mut Self::Transcript
) -> Result<Self::PermutationChallenge, PolyIOPErrors>

Step 1 of the IOP. Generate challenge beta and gamma from a transcript.

fn update_challenge(
challenge: &mut Self::PermutationChallenge,
transcript: &mut Self::Transcript,
prod_x_binding: &F
) -> Result<(), PolyIOPErrors>

Step 4 of the IOP. Update the challenge with alpha; returns an error if alpha already exists.

fn compute_products(
challenge: &Self::PermutationChallenge,
fx: &DenseMultilinearExtension<F>,
gx: &DenseMultilinearExtension<F>,
s_perm: &DenseMultilinearExtension<F>
) -> Result<[DenseMultilinearExtension<F>; 7], PolyIOPErrors>

Step 2 of the IOP. Compute the following 7 polynomials

prod(x)

prod(0, x)

prod(1, x)

prod(x, 0)

prod(x, 1)

numerator

denominator

where

prod(0,x) := prod(0, x0, x1, …, xn) which is the MLE over the evaluations of the following polynomial on the boolean hypercube {0,1}^n:

(f(x) + \beta s_id(x) + \gamma)/(g(x) + \beta s_perm(x) + \gamma)

where

beta and gamma are challenges

f(x), g(x), s_id(x), s_perm(x) are mle-s

prod(1,x) := prod(x, 0) * prod(x, 1)

numerator is the MLE for f(x) + \beta s_id(x) + \gamma

denominator is the MLE for g(x) + \beta s_perm(x) + \gamma

The caller needs to check num_vars matches in f/g/s_id/s_perm Cost: linear in N.

fn prove(
prod_x_and_aux_info: &[DenseMultilinearExtension<F>; 7],
challenge: &Self::PermutationChallenge,
transcript: &mut IOPTranscript<F>
) -> Result<Self::Proof, PolyIOPErrors>

Step 5 of the IOP.

Initialize the prover to argue that an MLE g(x) is a permutation of MLE f(x) over a permutation given by s_perm. Inputs:

7 MLEs from Self::compute_products

challenge: Self::Challenge that has been updated

transcript: a transcript that is used to generate the challenges beta and gamma Cost: O(N)

fn verify(
proof: &Self::Proof,
aux_info: &Self::VPAuxInfo,
transcript: &mut Self::Transcript
) -> Result<Self::PermutationCheckSubClaim, PolyIOPErrors>

Verify that an MLE g(x) is a permutation of MLE f(x) over a permutation given by s_perm.