Trait poly_iop::ProductCheck

pub trait ProductCheck<F: PrimeField>: ZeroCheck<F> {
    type ProductCheckSubClaim;
    type ProductCheckChallenge;
    type ProductProof;

    fn init_transcript() -> Self::Transcript;
    fn generate_challenge(
        transcript: &mut Self::Transcript
    ) -> Result<Self::ProductCheckChallenge, PolyIOPErrors>;
    fn compute_product_poly(
        fx: &VirtualPolynomial<F>
    ) -> Result<DenseMultilinearExtension<F>, PolyIOPErrors>;
    fn prove(
        fx: &VirtualPolynomial<F>, 
        prod_x: &DenseMultilinearExtension<F>, 
        transcript: &mut IOPTranscript<F>, 
        claimed_product: F
    ) -> Result<Self::ProductProof, PolyIOPErrors>;
    fn verify(
        proof: &Self::ProductProof, 
        aux_info: &Self::VPAuxInfo, 
        transcript: &mut Self::Transcript, 
        claimed_product: F
    ) -> Result<Self::ProductCheckSubClaim, PolyIOPErrors>;
}

A ProductCheck is derived from ZeroCheck.

A ProductCheck IOP takes the following steps:

Inputs:

f(x)

Prover steps:

compute_product_poly to build prod(x0, ..., x_n) from virtual polynomial f

push commitments of f(x), prod(x) to the transcript (done by the snark caller)

generate_challenge from current transcript (generate alpha)

prove to generate the zerocheck proof for the virtual polynomial prod(1, x) - prod(x, 0) * prod(x, 1) + alpha * (f(x) - prod(0, x))

Verifier steps:

Extract commitments of f(x), prod(x) from the proof, push them to the transcript (done by the snark caller)

generate_challenge from current transcript (generate alpha)

verify to verify the zerocheck proof and generate the subclaim for polynomial evaluations

Required Associated Types

type ProductCheckSubClaim

type ProductCheckChallenge

type ProductProof

Required Methods

fn init_transcript() -> Self::Transcript

Initialize the system with a transcript

This function is optional – in the case where a ProductCheck is an building block for a more complex protocol, the transcript may be initialized by this complex protocol, and passed to the ProductCheck prover/verifier.

fn generate_challenge(
transcript: &mut Self::Transcript
) -> Result<Self::ProductCheckChallenge, PolyIOPErrors>

Generate random challenge alpha from a transcript.

fn compute_product_poly(
fx: &VirtualPolynomial<F>
) -> Result<DenseMultilinearExtension<F>, PolyIOPErrors>

Compute the product polynomial prod(x) where

prod(0,x) := prod(0, x1, …, xn) is the MLE over the evaluations of f(x) on the boolean hypercube {0,1}^n

prod(1,x) is a MLE over the evaluations of prod(x, 0) * prod(x, 1) on the boolean hypercube {0,1}^n

The caller needs to check num_vars matches in f Cost: linear in N.

fn prove(
fx: &VirtualPolynomial<F>,
prod_x: &DenseMultilinearExtension<F>,
transcript: &mut IOPTranscript<F>,
claimed_product: F
) -> Result<Self::ProductProof, PolyIOPErrors>

Initialize the prover to argue that for a virtual polynomial f(x), it holds that s = \prod_{x \in {0,1}^n} f(x)

Inputs:

fx: the virtual polynomial

prod_x: the product polynomial

transcript: a transcript that is used to generate the challenges alpha

claimed_product: the claimed product value

Cost: O(N)

fn verify(
proof: &Self::ProductProof,
aux_info: &Self::VPAuxInfo,
transcript: &mut Self::Transcript,
claimed_product: F
) -> Result<Self::ProductCheckSubClaim, PolyIOPErrors>

Verify that for a witness virtual polynomial f(x), it holds that s = \prod_{x \in {0,1}^n} f(x)