Trait poly_iop::ZeroCheck

pub trait ZeroCheck<F: PrimeField>: SumCheck<F> {
    type ZeroCheckSubClaim;
    type ZeroCheckProof;

    fn init_transcript() -> Self::Transcript;
    fn prove(
        poly: &Self::VirtualPolynomial, 
        transcript: &mut Self::Transcript
    ) -> Result<Self::ZeroCheckProof, PolyIOPErrors>;
    fn verify(
        proof: &Self::ZeroCheckProof, 
        aux_info: &Self::VPAuxInfo, 
        transcript: &mut Self::Transcript
    ) -> Result<Self::ZeroCheckSubClaim, PolyIOPErrors>;
}

A ZeroCheck is derived from SumCheck.

Required Associated Types

type ZeroCheckSubClaim

type ZeroCheckProof

Required Methods

fn init_transcript() -> Self::Transcript

Initialize the system with a transcript

This function is optional – in the case where a ZeroCheck is an building block for a more complex protocol, the transcript may be initialized by this complex protocol, and passed to the ZeroCheck prover/verifier.

fn prove(
poly: &Self::VirtualPolynomial,
transcript: &mut Self::Transcript
) -> Result<Self::ZeroCheckProof, PolyIOPErrors>

initialize the prover to argue for the sum of polynomial over {0,1}^num_vars is zero.

fn verify(
proof: &Self::ZeroCheckProof,
aux_info: &Self::VPAuxInfo,
transcript: &mut Self::Transcript
) -> Result<Self::ZeroCheckSubClaim, PolyIOPErrors>

verify the claimed sum using the proof